+++ one/tests/test_one.py
@@ -1101,7 +1101,8 @@ class TestOneAlyx(unittest.TestCase):
         self.assertEqual(len(out_paths), N, 'Returned list should equal length of input datasets')
         self.assertTrue(all(isinstance(x, Path) for x in out_paths))
         # These values come from REST cache fixture
-        boto3_mock.assert_called_with(aws_access_key_id='ABCDEF', aws_secret_access_key='shhh')
+        boto3_mock.assert_called_with(aws_access_key_id='ABCDEF', aws_secret_access_key='shhh',
+                                      region_name=None)
         ((bucket, path), _), *_ = boto3_mock().resource().Object.call_args_list
         self.assertEqual(bucket, 's3_bucket')
         self.assertTrue(dsets['rel_path'][0].split('.')[0] in path)
@@ -1391,6 +1392,10 @@ class TestOneSetup(unittest.TestCase):
         self.tempdir = tempfile.TemporaryDirectory()
         self.addCleanup(self.tempdir.cleanup)
         self.get_file = partial(util.get_file, self.tempdir.name)
+        # Change default cache dir to temporary directory
+        patch = mock.patch('one.params.CACHE_DIR_DEFAULT', Path(self.tempdir.name))
+        patch.start()
+        self.addCleanup(patch.stop)
 
     def test_local_cache_setup_prompt(self):
         """Test One.setup"""
@@ -1499,9 +1504,7 @@ class TestOneSetup(unittest.TestCase):
     def test_patch_params(self):
         """Test patching legacy params to the new location"""
         # Save some old-style params
-        old_pars = (one.params.default()
-                    .set('CACHE_DIR', self.tempdir.name)
-                    .set('HTTP_DATA_SERVER', 'openalyx.org'))
+        old_pars = one.params.default().set('HTTP_DATA_SERVER', 'openalyx.org')
         # Save a REST query in the old location
         old_rest = Path(self.tempdir.name, '.one', '.rest', old_pars.ALYX_URL[8:], 'https')
         old_rest.mkdir(parents=True, exist_ok=True)