+++ CHANGELOG.md
@@ -12,6 +12,9 @@
 - added clarification for generating cache with an Alyx database in documentation
 - REST cache now stored in download cache directory
 - JSON tests may now be run concurrently
+- fix'd dimension check for DataFrame attributes in ALF objects
+- dimension warning logged when dimensions don't match after appending to ALFBunch
+- created_time now updated correctly in cache meta
 
 ## [1.11.0]